added changes to contract.js and updated readme
