+++ stellar/README.md
@@ -280,37 +280,44 @@ node stellar/gmp.js execute avalanche '0x0bcbbfc9b006db6958f3fce75f11fdc306b45e8
 
 ## Pausable Contract Functionality
 
-This script allows you to manage the pausable state of a Stellar contract. You can check if the contract is paused, pause the contract, or unpause the contract.
+Here is how to manage the pausable state of a Stellar contract. You can check if the contract is paused, pause the contract, or unpause the contract.
 
 #### Usage
 
 To use this script, run the following command with the appropriate options:
 
 ```bash
-node stellar/pause.js --action <action> --address <contract-address>
+node stellar/contract.js [action] [contract-name]
 ```
+
+
 
 #### Options
 
-- `--action <action>`: The action to perform on the contract. Choices are `paused`, `pause`, and `unpause`. This option is mandatory.
-- `--address <contract-address>`: The address of the contract to interact with. This option is mandatory.
+- `[action]` can be one of the following:
+  - `pause`: Pause the contract
+  - `paused`: Check if the contract is paused
+  - `unpause`: Unpause the contract
+
+- `[contract-name]`: The name of the contract to interact with. This option is mandatory.
+
 
 #### Examples
 
 Check if the contract is paused:
 
 ```bash
-node stellar/pause.js --action paused --address CBC2TFJUF4PIZFZXRJJ3K47OBZW5BUJVWBELZDJGMNGNVVNA4A64D5FI
+node stellar/contract.js paused axelar_gateway
 ```
 
 Pause the contract:
 
 ```bash
-node stellar/pause.js --action pause --address CBC2TFJUF4PIZFZXRJJ3K47OBZW5BUJVWBELZDJGMNGNVVNA4A64D5FI
+node stellar/contract.js pause axelar_gateway
 ```
 
 Unpause the contract:
 
 ```bash
-node stellar/pause.js --action unpause --address CBC2TFJUF4PIZFZXRJJ3K47OBZW5BUJVWBELZDJGMNGNVVNA4A64D5FI
+node stellar/contract.js unpause axelar_gateway
 ```
